+++ sources/core/objects/Tile.h
@@ -10,7 +10,7 @@
 template<class T>
 FastSerializer& operator<<(FastSerializer& file, const std::vector<IdPtr<T>>& content)
 {
-    file << content.size();
+    file << static_cast<quint32>(content.size());
     for (auto it = content.begin(); it != content.end(); ++it)
     {
         file << it->Id();
